Refactor module options and add a store interface
